+++ admin/regroup/RegroupDialogModel.js
@@ -35,8 +35,8 @@ export class RegroupDialogModel extends HoistModel {
 
     async saveAsync() {
         const {_parent, groupName} = this,
-            {selection, store} = _parent.gridModel,
-            ids = selection.map(it => it.id),
+            {selectedRecords, store} = _parent.gridModel,
+            ids = selectedRecords.map(it => it.id),
             resp = await store.bulkUpdateRecordsAsync(ids, {groupName}),
             failuresPresent = resp.fail > 0,
             intent = failuresPresent ? 'warning' : 'success';