+++ cmp/grid/GridModel.js
@@ -23,7 +23,8 @@ import {
     throwIf,
     warnIf,
     withDebug,
-    withDefault
+    withDefault,
+    apiDeprecated
 } from '@xh/hoist/utils/js';
 import equal from 'fast-deep-equal';
 import {
@@ -542,11 +543,11 @@ export class GridModel extends HoistModel {
         if (!isReady) return;
 
         const {agApi, selModel} = this,
-            {records} = selModel,
+            {selectedRecords} = selModel,
             indices = [];
 
         // 1) Expand any selected nodes that are collapsed
-        records.forEach(({id}) => {
+        selectedRecords.forEach(({id}) => {
             for (let row = agApi.getRowNode(id)?.parent; row; row = row.parent) {
                 if (!row.expanded) {
                     agApi.setRowNodeExpanded(row, true);
@@ -557,13 +558,13 @@ export class GridModel extends HoistModel {
         await wait();
 
         // 2) Scroll to all selected nodes
-        records.forEach(({id}) => {
+        selectedRecords.forEach(({id}) => {
             const rowIndex = agApi.getRowNode(id)?.rowIndex;
             if (!isNil(rowIndex)) indices.push(rowIndex);
         });
 
         const indexCount = indices.length;
-        if (indexCount !== records.length) {
+        if (indexCount !== selectedRecords.length) {
             console.warn('Grid row nodes not found for all selected records.');
         }
 
@@ -578,17 +579,20 @@ export class GridModel extends HoistModel {
     /** @return {boolean} - true if any records are selected. */
     get hasSelection() {return !this.selModel.isEmpty}
 
-    /** @return {Record[]} - currently selected Records. */
-    get selection() {return this.selModel.records}
+    /** @return {Record[]} - currently selected records. */
+    get selectedRecords() {return this.selModel.selectedRecords}
+
+    /** @return {RecordId[]} - IDs of currently selected records. */
+    get selectedIds() {return this.selModel.selectedIds}
 
     /**
      * @return {?Record} - single selected record, or null if multiple/no records selected.
      *
      * Note that this getter will also change if just the data of selected record is changed
      * due to store loading or editing.  Applications only interested in the identity
-     * of the selection should use {@see selectedRecordId} instead.
+     * of the selection should use {@see selectedId} instead.
      */
-    get selectedRecord() {return this.selModel.singleRecord}
+    get selectedRecord() {return this.selModel.selectedRecord}
 
     /**
      * @return {?RecordId} - ID of selected record, or null if multiple/no records selected.
@@ -597,7 +601,7 @@ export class GridModel extends HoistModel {
      * due to store loading or editing.  Applications also interested in the contents of the
      * of the selection should use the {@see selectedRecord} getter instead.
      */
-    get selectedRecordId() {return this.selModel.selectedRecordId}
+    get selectedId() {return this.selModel.selectedId}
 
     /** @return {boolean} - true if this grid has no records to show in its store. */
     get empty() {return this.store.empty}
@@ -1031,16 +1035,16 @@ export class GridModel extends HoistModel {
         const isReady = await this.whenReadyAsync();
         if (!isReady) return;
 
-        const {store, agGridModel, agApi, selection} = this;
+        const {store, agGridModel, agApi, selectedRecords} = this;
 
         let recToEdit;
         if (record) {
             // Normalize specified record, if any.
             recToEdit = record.isRecord ? record : store.getById(record);
         } else {
-            if (!isEmpty(selection)) {
+            if (!isEmpty(selectedRecords)) {
                 // Or use first selected record, if any.
-                recToEdit = selection[0];
+                recToEdit = selectedRecords[0];
             } else {
                 // Or use the first record overall.
                 const firstRowId = agGridModel.getFirstSelectableRowNodeId();
@@ -1123,6 +1127,18 @@ export class GridModel extends HoistModel {
         }
 
         return this.isReady;
+    }
+
+    /** @deprecated */
+    get selection() {
+        apiDeprecated('selection', {msg: 'Use selectedRecords instead', v: 'v44'});
+        return this.selectedRecords;
+    }
+
+    /** @deprecated */
+    get selectedRecordId() {
+        apiDeprecated('selectedRecordId', {msg: 'Use selectedId instead', v: 'v44'});
+        return this.selectedId;
     }
 
     //-----------------------