+++ data/StoreSelectionModel.js
@@ -8,6 +8,7 @@
 import {HoistModel} from '@xh/hoist/core';
 import {action, computed, observable, makeObservable} from '@xh/hoist/mobx';
 import {castArray, compact, remove, isEqual, union, map} from 'lodash';
+import {apiDeprecated} from '@xh/hoist/utils/js';
 
 /**
  * Model for managing store selections.
@@ -41,14 +42,14 @@ export class StoreSelectionModel extends HoistModel {
 
     /** @return {Record[]} - currently selected records. */
     @computed.struct
-    get records() {
+    get selectedRecords() {
         return compact(this._ids.map(it => this.store.getById(it, true)));
     }
 
     /** @return {RecordId[]} - IDs of currently selected records. */
     @computed.struct
-    get ids() {
-        return map(this.records, 'id');
+    get selectedIds() {
+        return map(this.selectedRecords, 'id');
     }
 
     /**
@@ -56,11 +57,11 @@ export class StoreSelectionModel extends HoistModel {
      *
      * Note that this getter will also change if just the data of selected record is changed
      * due to store loading or editing.  Applications only interested in the identity
-     * of the selection should use {@see selectedRecordId} instead.
+     * of the selection should use {@see selectedId} instead.
      */
-    get singleRecord() {
-        const {records} = this;
-        return records.length === 1 ? records[0] : null;
+    get selectedRecord() {
+        const {selectedRecords} = this;
+        return selectedRecords.length === 1 ? selectedRecords[0] : null;
     }
 
     /**
@@ -68,11 +69,11 @@ export class StoreSelectionModel extends HoistModel {
      *
      * Note that this getter will *not* change if just the data of selected record is changed
      * due to store loading or editing.  Applications also interested in the contents of the
-     * of the selection should use the {@see singleRecord} getter instead.
+     * of the selection should use the {@see selectedRecord} getter instead.
      */
-    get selectedRecordId() {
-        const {ids} = this;
-        return ids.length === 1 ? ids[0] : null;
+    get selectedId() {
+        const {selectedIds} = this;
+        return selectedIds.length === 1 ? selectedIds[0] : null;
     }
 
     /** @return {boolean} - true if selection is empty. */
@@ -83,7 +84,7 @@ export class StoreSelectionModel extends HoistModel {
     /** @return {number} - count of currently selected records. */
     @computed
     get count() {
-        return this.records.length;
+        return this.selectedRecords.length;
     }
 
     /**
@@ -119,11 +120,34 @@ export class StoreSelectionModel extends HoistModel {
         }
     }
 
-
     /** Clear the selection. */
     @action
     clear() {
         this.select([]);
+    }
+
+    /** @deprecated */
+    get records() {
+        apiDeprecated('records', {msg: 'Use selectedRecords instead', v: 'v44'});
+        return this.selectedRecords;
+    }
+
+    /** @deprecated */
+    get ids() {
+        apiDeprecated('ids', {msg: 'Use selectedIds instead', v: 'v44'});
+        return this.selectedIds;
+    }
+
+    /** @deprecated */
+    get singleRecord() {
+        apiDeprecated('singleRecord', {msg: 'Use selectedRecord instead', v: 'v44'});
+        return this.selectedRecord;
+    }
+
+    /** @deprecated */
+    get selectedRecordId() {
+        apiDeprecated('selectedRecordId', {msg: 'Use selectedId instead', v: 'v44'});
+        return this.selectedId;
     }
 
     //------------------------